requirements.txt
# Core
python-dotenv>=1.0
pyyaml>=6.0

# Deep Learning
torch>=2.0
torchvision>=0.15
timm>=0.9

# Geospatial
rasterio>=1.3
geopandas>=0.14
shapely>=2.0
pyproj>=3.6
folium>=0.15

# Dashboard
streamlit>=1.28
streamlit-folium>=0.15
plotly>=5.18

# Data
numpy>=1.24
pandas>=2.0
scikit-learn>=1.3
Pillow>=10.0
matplotlib>=3.8
seaborn>=0.13
tqdm>=4.66

# Dev
pytest>=7.0
pytest-cov>=4.0
ruff>=0.1
pre-commit>=3.0
